Modularized user input form and converted data into separate components
